+++ tests/Integration/Syntax/v5/ShouldBeOfType.v5.tests.ps1
@@ -12,7 +12,6 @@ Describe 'Should -BeOfType' {
             'AnyString' | Should -BeOfType [System.String] 'BecauseString'
         }
 
-
         It 'Should convert `Should -ActualValue ''AnyString'' [System.String] -BeOfType ''BecauseString''` correctly' {
             Should -ActualValue 'AnyString' [System.String] -BeOfType 'BecauseString'
         }
@@ -21,20 +20,20 @@ Describe 'Should -BeOfType' {
             Should -ActualValue 'AnyString' [System.String] -BeOfType 'BecauseString'
         }
 
-        It 'Should convert `Should -ActualValue ''AnyString'' -ExpectedValue [System.String] -BeOfType -Because ''BecauseString''` correctly' {
-            Should -ActualValue 'AnyString' -ExpectedValue [System.String] -BeOfType -Because 'BecauseString'
+        It 'Should convert `Should -ActualValue ''AnyString'' -ExpectedType [System.String] -BeOfType -Because ''BecauseString''` correctly' {
+            Should -ActualValue 'AnyString' -ExpectedType [System.String] -BeOfType -Because 'BecauseString'
         }
 
         It 'Should convert `Should -BeOfType [System.String] -ActualValue ''AnyString''` correctly' {
             Should -BeOfType [System.String] -ActualValue 'AnyString'
         }
 
-        It 'Should convert `Should -ExpectedValue [System.String] -ActualValue ''AnyString'' -BeOfType` correctly' {
-            Should -ExpectedValue [System.String] -ActualValue 'AnyString' -BeOfType
+        It 'Should convert `Should -ExpectedType [System.String] -ActualValue ''AnyString'' -BeOfType` correctly' {
+            Should -ExpectedType [System.String] -ActualValue 'AnyString' -BeOfType
         }
 
-        It 'Should convert `Should -Not:$false -ExpectedValue [System.String] -ActualValue ''AnyString'' -BeOfType` correctly' {
-            Should -Not:$false -ExpectedValue [System.String] -ActualValue 'AnyString' -BeOfType
+        It 'Should convert `Should -Not:$false -ExpectedType [System.String] -ActualValue ''AnyString'' -BeOfType` correctly' {
+            Should -Not:$false -ExpectedType [System.String] -ActualValue 'AnyString' -BeOfType
         }
 
         It 'Should convert `Should -BeOfType (Get-Something)` correctly' {
@@ -77,18 +76,18 @@ Describe 'Should -BeOfType' {
 
     Context 'When tests should always use positional parameters' {
         Context 'When the tests are affirming' {
-            It 'Should convert `Should -HaveType -ExpectedValue ([System.String]) -ActualValue ''AnyString''` correctly' {
-                Should -BeOfType -ExpectedValue ([System.String]) -ActualValue 'AnyString'
+            It 'Should convert `Should -BeOfType -ExpectedType ([System.String]) -ActualValue ''AnyString''` correctly' {
+                Should -BeOfType -ExpectedType ([System.String]) -ActualValue 'AnyString'
             }
 
-            It 'Should convert `Should -HaveType -ExpectedValue ([System.String]) -ActualValue ''AnyString'' -Because ''BecauseString''` correctly' {
-                Should -BeOfType -ExpectedValue ([System.String]) -ActualValue 'AnyString'
+            It 'Should convert `Should -BeOfType -ExpectedType ([System.String]) -ActualValue ''AnyString'' -Because ''BecauseString''` correctly' {
+                Should -BeOfType -ExpectedType ([System.String]) -ActualValue 'AnyString'
             }
         }
 
         Context 'When the tests are negated' {
-            It 'Should convert `Should -HaveType -ExpectedValue ([System.String]) -Not -ActualValue ''AnyString'' -Because ''BecauseString''` correctly' {
-                Should -BeOfType -ExpectedValue ([System.String]) -Not -ActualValue 2 -Because 'BecauseString'
+            It 'Should convert `Should -BeOfType -ExpectedType ([System.String]) -Not -ActualValue ''AnyString'' -Because ''BecauseString''` correctly' {
+                Should -BeOfType -ExpectedType ([System.String]) -Not -ActualValue 2 -Because 'BecauseString'
             }
         }
     }