+++ CHANGELOG.md
@@ -45,3 +45,4 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   to correctly handle scenario when actual value and expected value are the same.
 - Negated `Should -Not -BeGreaterThan` now converts to `Should-BeLessThanOrEqual`
   to correctly handle scenario when actual value and expected value are the same.
+- Fix parameter name in `Convert-ShouldBeOfType`